fix clippy lints
